refactor: flatten domain folders
