+++ src/domains/tasks/Offers/OfferStyles.tsx
