+++ src/domains/tasks/TaskDetails/OffersList.tsx
