+++ src/domains/tasks/TaskDetails/TaskDetails.tsx
@@ -13,9 +13,7 @@ import {
   faHourglassStart,
 } from "@fortawesome/free-solid-svg-icons";
 import Image from "next/legacy/image";
-import useForm from "../../../../../lib/useForm";
 import { OffersList } from "./OffersList";
-import { validate } from "../../../../../lib/validator";
 
 import Link from "next/link";
 import ErrorMessage from "@Components/common/ErrorMessage";
@@ -24,6 +22,8 @@ import Loading from "@Components/common/Loading";
 import ButtonStyles from "@Components/styles/ButtonStyles";
 import FormStyles from "@Components/styles/FormStyles";
 import SubmitFormStyles from "@Components/styles/SubmitFormStyles";
+import useForm from "src/lib/useForm";
+import { validate } from "src/lib/validator";
 
 export function TaskDetails({ selectedTask }: any) {
   // const { selectedTask } = useContext(DashboardContext);
