+++ src/pages/mytasks/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import TasksDashboard from "@Tasks/features/Tasks/TasksDashboard";
+import TasksDashboard from "@Tasks/TasksDashboard";
 
 const MyTasksPage: NextPage = () => {
   return <TasksDashboard myTasks={true} />;
