+++ src/pages/offers.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import { Dashboard } from "@Tasks/features/Offers";
+import { Dashboard } from "@Tasks/Offers";
 
 const OffersPage: NextPage = () => {
   return <Dashboard />;
