+++ src/pages/tasks/[id].tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
 import { useRouter } from "next/router";
-import TasksDashboard from "@Tasks/features/Tasks/TasksDashboard";
+import TasksDashboard from "@Tasks/TasksDashboard";
 
 const TasksPage: NextPage = () => {
   const router = useRouter();
