+++ src/pages/tasks/index.tsx
@@ -1,4 +1,4 @@
-import TasksDashboard from "@Tasks/features/Tasks/TasksDashboard";
+import TasksDashboard from "@Tasks/TasksDashboard";
 
 const TasksPage = () => {
   return <TasksDashboard myTasks={false} />;
